+++ MLEM.Ui/Elements/ElementHelper.cs
@@ -2,6 +2,7 @@ using System;
 using System.Linq;
 using Microsoft.Xna.Framework;
 using MLEM.Input;
+using MLEM.Misc;
 using MLEM.Textures;
 
 namespace MLEM.Ui.Elements {
@@ -22,7 +23,8 @@ namespace MLEM.Ui.Elements {
         /// <returns>An image button</returns>
         public static Button ImageButton(Anchor anchor, Vector2 size, TextureRegion texture, string text = null, string tooltipText = null, float imagePadding = 2) {
             var button = new Button(anchor, size, text, tooltipText);
-            var image = new Image(Anchor.CenterLeft, Vector2.One, texture) {Padding = new Vector2(imagePadding)};
+            var image = new Image(Anchor.CenterLeft, Vector2.One, texture);
+            image.Padding.SetFromStyle(new Padding(imagePadding), 1);
             button.OnAreaUpdated += e => image.Size = new Vector2(e.Area.Height, e.Area.Height) / e.Scale;
             button.AddChild(image, 0);
             return button;