+++ Tests/UiTests.cs
@@ -49,7 +49,7 @@ namespace Tests {
                 var button = panel.AddChild(new Button(Anchor.AutoLeft, new Vector2(1)) {
                     SetHeightBasedOnChildren = true,
                     Padding = new Padding(0, 0, 0, 1),
-                    ChildPadding = new Vector2(3)
+                    ChildPadding = new Padding(3)
                 });
                 button.AddChild(new Group(Anchor.AutoLeft, new Vector2(0.5F, 30), false) {
                     CanBeMoused = false